extract update control methods into separate files
