+++ server/src/js/components/bar-chart/index.mjs
@@ -7,6 +7,7 @@ import { timeFormat } from 'd3-time-format'
 import { select, pointer } from 'd3-selection'
 import { max } from 'd3-array'
 import { timeMinute } from 'd3-time'
+import { updatePagination, updateSegmentedControl } from './controls.mjs'
 // const { xhr } = window.flood.utils
 const { forEach } = window.flood.utils
 
@@ -237,52 +238,6 @@ function BarChart (containerId, stationId, data) {
     locatorLine.classed('locator__line--visible', false)
     locatorBackground.classed('locator__background--visible', false)
   }
-
-  const updateSegmentedControl = () => {
-    const now = new Date()
-    const dataDurationDays = (new Date(now.getTime() - dataStart.getTime())) / (1000 * 60 * 60 * 24)
-    // Check there are at least 2 telemetry arrays
-    let numBands = 0
-    for (let i = 0; i < bands.length; i++) {
-      numBands += Object.getOwnPropertyDescriptor(dataCache, bands[i].period) ? 1 : 0
-    }
-    // Determine which controls to display
-    forEach(segmentedControl.querySelectorAll('.defra-chart-segmented-control input'), input => {
-      const isBand = period === input.getAttribute('data-period')
-      const band = bands.find(x => x.period === input.getAttribute('data-period'))
-      input.checked = isBand
-      input.parentNode.style.display = (band.days <= dataDurationDays) && numBands > 1 ? 'inline-block' : 'none'
-      input.parentNode.classList.toggle('defra-chart-segmented-control__segment--selected', isBand)
-    })
-  }
-
-  const updatePagination = (start, end, duration, durationHours) => {
-    // Set paging values and ensure they are within data range
-    const now = new Date()
-    let nextStart = new Date(start.getTime() + duration)
-    let nextEnd = new Date(end.getTime() + duration)
-    let previousStart = new Date(start.getTime() - duration)
-    let previousEnd = new Date(end.getTime() - duration)
-    nextEnd = nextEnd.getTime() <= now.getTime() ? nextEnd.toISOString().replace(/.\d+Z$/g, 'Z') : null
-    nextStart = nextEnd ? nextStart.toISOString().replace(/.\d+Z$/g, 'Z') : null
-    previousStart = previousStart.getTime() >= dataStart.getTime() ? previousStart.toISOString().replace(/.\d+Z$/g, 'Z') : null
-    previousEnd = previousStart ? previousEnd.toISOString().replace(/.\d+Z$/g, 'Z') : null
-    // Set properties
-    paginationInner.style.display = (nextStart || previousEnd) ? 'inline-block' : 'none'
-    pageForward.setAttribute('data-start', nextStart)
-    pageForward.setAttribute('data-end', nextEnd)
-    pageBack.setAttribute('data-start', previousStart)
-    pageBack.setAttribute('data-end', previousEnd)
-    pageForward.setAttribute('aria-disabled', !(nextStart && nextEnd))
-    pageBack.setAttribute('data-journey-click', 'Rainfall:Chart Interaction:Rainfall - Previous 24hrs')
-    pageForward.setAttribute('data-journey-click', 'Rainfall:Chart Interaction:Rainfall - Next 24hrs')
-    pageBack.setAttribute('aria-disabled', !(previousStart && previousEnd))
-    pageForwardText.innerText = `Next ${durationHours > 1 ? durationHours : duration} ${durationHours > 1 ? 'hours' : 'minutes'}`
-    pageBackText.innerText = `Previous ${durationHours > 1 ? durationHours : duration} ${durationHours > 1 ? 'hours' : 'minutes'}`
-    pageForwardDescription.innerText = ''
-    pageBackDescription.innerText = ''
-  }
-
   const updateGrid = (colcount, total, hours, days, start, end) => {
     // Update grid properites
     grid.attr('aria-rowcount', 1)
@@ -344,8 +299,21 @@ function BarChart (containerId, stationId, data) {
       dataItem = direction === 'forward' ? dataPage[positiveDataItems[positiveDataItems.length - 1]] : dataPage[positiveDataItems[0]]
     }
     // Update html control properties
-    updateSegmentedControl()
-    updatePagination(pageStart, pageEnd, pageDuration, pageDurationHours)
+    updateSegmentedControl({ bands, dataCache, dataStart, period, segmentedControl })
+    updatePagination({
+      start: pageStart,
+      end: pageEnd,
+      duration: pageDuration,
+      durationHours: pageDurationHours,
+      dataStart,
+      paginationInner,
+      pageForward,
+      pageForwardText,
+      pageForwardDescription,
+      pageBack,
+      pageBackText,
+      pageBackDescription
+    })
     const totalPageRainfall = dataPage.reduce((a, b) => { return a + b.value }, 0)
     const pageValueStart = new Date(new Date(dataPage[dataPage.length - 1].dateTime).getTime() - valueDuration)
     const pageValueEnd = new Date(dataPage[0].dateTime)
